modified:   Zombie Swarm/Bullet.java 	modified:   Zombie Swarm/GamePanel.java 	modified:   Zombie Swarm/Hero.java 	modified:   Zombie Swarm/gun.java 	modified:   Zombie Swarm/Gun.java

+++ Zombie Swarm/Hero.java
@@ -148,7 +148,7 @@ public class Hero extends GameObject {
         return gun.getAmmo();
     }
 
-    public int getFireRate(){
+    public double getFireRate(){
         return gun.getFireRate();
     }
